Exception for email
Do not verify that the email you enter is an email in HTML.
Just for admin panel and for login.
The idea is to be able to use a non email ID for visitor (ocb-oops). You cannot create an account without an email but the admin can change an email/ID to normal text